+++ crates/wreq-util/tests/emulate_chrome.rs
@@ -275,3 +275,23 @@ test_emulation!(
     ["t13d1516h2_8daaf6152771_806a8c22fdea"],
     "52d84b11737d980aef856699f885ca86"
 );
+
+test_emulation!(
+    test_chrome152,
+    Emulation::Chrome152,
+    [
+        "t13d1517h2_8daaf6152771_cb7bf5808d99",
+        "t13d1518h2_8daaf6152771_e2d80978ab2e"
+    ],
+    "52d84b11737d980aef856699f885ca86"
+);
+
+test_emulation!(
+    test_chrome153,
+    Emulation::Chrome153,
+    [
+        "t13d1517h2_8daaf6152771_cb7bf5808d99",
+        "t13d1518h2_8daaf6152771_e2d80978ab2e"
+    ],
+    "52d84b11737d980aef856699f885ca86"
+);