+++ test/emulation_test.rb
@@ -18,6 +18,16 @@ class EmulationTest < Minitest::Test
     assert_instance_of Wreq::Emulation, Wreq::Emulation.new(profile: profile)
   end
 
+  def test_new_chrome_profiles_are_available
+    {
+      Wreq::Profile::Chrome152 => "Chrome152",
+      Wreq::Profile::Chrome153 => "Chrome153"
+    }.each do |profile, name|
+      assert_equal name, profile.to_s
+      assert_instance_of Wreq::Emulation, Wreq::Emulation.new(profile: profile)
+    end
+  end
+
   def test_all_emulation_os_constants_are_non_nil
     Wreq::Platform.constants.each do |name|
       const = Wreq::Platform.const_get(name)